+++ src/main/java/br/com/process/integration/database/model/entity/service/example/EntitySixService.java
@@ -1,17 +1,19 @@
 package br.com.process.integration.database.model.entity.service.example;
 
-import br.com.process.integration.database.core.application.AbstractJpaService;
-import br.com.process.integration.database.core.ui.QueryJpaController;
-import br.com.process.integration.database.model.entity.dto.example.EntitySix;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.web.PagedResourcesAssembler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import br.com.process.integration.database.core.application.AbstractJpaService;
+import br.com.process.integration.database.core.ui.QueryJpaController;
+import br.com.process.integration.database.model.entity.dto.example.EntitySix;
+
 @Service
 @Transactional
-public class EntitySixService extends AbstractJpaService<EntitySix, EntitySix, Integer> {
+public class EntitySixService extends AbstractJpaService<EntitySix, EntitySix, Long> {
+	
     @Autowired
     private PagedResourcesAssembler<EntitySix> pagedResourcesAssembler;
 
@@ -27,7 +29,7 @@ public class EntitySixService extends AbstractJpaService<EntitySix, EntitySix, I
     }
 
     @Override
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
